backend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.1.7
django-autocomplete-light==3.9.7
django-cors-headers==3.14.0
django-filter==22.1
django-restframework==0.0.1
django-taggit==3.1.0
django-treebeard==4.7
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
